RegisterLessonInformationViewModel.swift에 선언되어있던 enum 분리
