+++ Nanagong/Nanagong/RegisterLessonView/Model/RegisterLessonInputForm.swift
@@ -1,0 +1,27 @@
+//
+//  RegisterLessonInputForm.swift
+//  Nanagong
+//
+//  Created by Olaf on 2021/10/30.
+//
+
+import Foundation
+
+enum RegisterInputFormType {
+    
+    case lessonName
+    
+    case numberOfLessons
+    
+    case lessonCategory
+    
+    case lessonSite
+    
+    case lessonStartDate
+    
+    case lessonEndDate
+    
+    case lessonPrice
+    
+    case lessonDetermination
+}
